feature: updateItem

+++ backend/project1/src/main/java/com/team4/project1/domain/item/dto/ItemDto.java
@@ -2,10 +2,11 @@ package com.team4.project1.domain.item.dto;
 
 import com.team4.project1.domain.item.entity.Item;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 public class ItemDto {
-    private Integer id;
+    private Long id;
     private String name;
     private Integer price;
 
@@ -16,7 +17,6 @@ public class ItemDto {
     }
 
     public static ItemDto from(Item item) {
-
         return new ItemDto(item);
     }
